# docs/swagger.json
{"consumes":["application/json"],"definitions":{"GetAllDependenciesResponseIdArray":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"GetBoardContentResponseBoardContent":{"properties":{"topics":{"items":{"$ref":"#/definitions/topicsTopicWithFetchedTasks","type":"object"},"type":"array"}},"type":"object"},"boardsBoard":{"properties":{"createdAt":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"string"},"name":{"type":"string"},"topicIds":{"items":{"format":"int64","type":"string"},"type":"array"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"boardsBoardOrder":{"properties":{"boardId":{"format":"int64","type":"string"},"place":{"format":"int32","type":"integer"}},"type":"object"},"boardsChangeBoardOrderRequest":{"properties":{"changes":{"items":{"$ref":"#/definitions/boardsBoardOrder","type":"object"},"type":"array"}},"type":"object"},"boardsChangeBoardOrderResponse":{"type":"object"},"boardsCreateBoardsRequest":{"properties":{"boardsToCreate":{"items":{"$ref":"#/definitions/boardsCreateBoardsRequestBoardPrototype","type":"object"},"type":"array"}},"type":"object"},"boardsCreateBoardsRequestBoardPrototype":{"properties":{"name":{"type":"string"},"topicIds":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsCreateBoardsResponse":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsDeleteBoardsRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsDeleteBoardsResponse":{"type":"object"},"boardsGetAllBoardsRequest":{"type":"object"},"boardsGetAllBoardsResponse":{"properties":{"boards":{"items":{"$ref":"#/definitions/boardsBoard","type":"object"},"type":"array"}},"type":"object"},"boardsGetBoardContentRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsGetBoardContentResponse":{"properties":{"contentById":{"additionalProperties":{"$ref":"#/definitions/GetBoardContentResponseBoardContent"},"type":"object"}},"type":"object"},"boardsGetBoardOrderRequest":{"type":"object"},"boardsGetBoardOrderResponse":{"properties":{"order":{"additionalProperties":{"format":"int32","type":"integer"},"type":"object"}},"type":"object"},"boardsGetBoardsRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsGetBoardsResponse":{"properties":{"boards":{"items":{"$ref":"#/definitions/boardsBoard","type":"object"},"type":"array"}},"type":"object"},"boardsUpdateBoardsRequest":{"properties":{"boardsToUpdate":{"items":{"$ref":"#/definitions/boardsUpdateBoardsRequestBoardPrototype","type":"object"},"type":"array"}},"type":"object"},"boardsUpdateBoardsRequestBoardPrototype":{"properties":{"id":{"format":"int64","type":"string"},"name":{"type":"string"},"topicIds":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"boardsUpdateBoardsResponse":{"type":"object"},"protobufAny":{"additionalProperties":{},"properties":{"@type":{"type":"string"}},"type":"object"},"rpcStatus":{"properties":{"code":{"format":"int32","type":"integer"},"details":{"items":{"$ref":"#/definitions/protobufAny","type":"object"},"type":"array"},"message":{"type":"string"}},"type":"object"},"tasksCreateTasksRequest":{"properties":{"tasksToCreate":{"items":{"$ref":"#/definitions/tasksCreateTasksRequestTaskPrototype","type":"object"},"type":"array"}},"type":"object"},"tasksCreateTasksRequestTaskPrototype":{"properties":{"deps":{"items":{"format":"int64","type":"string"},"type":"array"},"description":{"type":"string"},"finishTime":{"format":"date-time","type":"string"},"name":{"type":"string"},"priority":{"$ref":"#/definitions/tasksPriority"},"startTime":{"format":"date-time","type":"string"}},"type":"object"},"tasksCreateTasksResponse":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"tasksDeleteTasksRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"tasksDeleteTasksResponse":{"type":"object"},"tasksGetAllDependenciesRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"tasksGetAllDependenciesResponse":{"properties":{"dependenciesById":{"additionalProperties":{"$ref":"#/definitions/GetAllDependenciesResponseIdArray"},"type":"object"}},"type":"object"},"tasksGetTasksRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"tasksGetTasksResponse":{"properties":{"tasks":{"items":{"$ref":"#/definitions/tasksTask","type":"object"},"type":"array"}},"type":"object"},"tasksPriority":{"default":"PRIORITY_UNSPECIFIED","enum":["PRIORITY_UNSPECIFIED","PRIORITY_LOW","PRIORITY_MEDIUM","PRIORITY_HIGH","PRIORITY_CRITICAL"],"type":"string"},"tasksTask":{"properties":{"createdAt":{"format":"date-time","type":"string"},"deps":{"items":{"format":"int64","type":"string"},"type":"array"},"description":{"type":"string"},"finishTime":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"string"},"name":{"type":"string"},"priority":{"$ref":"#/definitions/tasksPriority"},"startTime":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"tasksUpdateTasksRequest":{"properties":{"tasksToUpdate":{"items":{"$ref":"#/definitions/tasksUpdateTasksRequestTaskPrototype","type":"object"},"type":"array"}},"type":"object"},"tasksUpdateTasksRequestTaskPrototype":{"properties":{"deps":{"items":{"format":"int64","type":"string"},"type":"array"},"description":{"type":"string"},"finishTime":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"string"},"name":{"type":"string"},"priority":{"$ref":"#/definitions/tasksPriority"},"startTime":{"format":"date-time","type":"string"}},"type":"object"},"tasksUpdateTasksResponse":{"type":"object"},"topicsCreateTopicsRequest":{"properties":{"topicsToCreate":{"items":{"$ref":"#/definitions/topicsCreateTopicsRequestTopicPrototype","type":"object"},"type":"array"}},"type":"object"},"topicsCreateTopicsRequestTopicPrototype":{"properties":{"name":{"type":"string"},"taskIds":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"topicsCreateTopicsResponse":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"topicsDeleteTopicsRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"topicsDeleteTopicsResponse":{"type":"object"},"topicsGetTopicsRequest":{"properties":{"ids":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"topicsGetTopicsResponse":{"properties":{"topics":{"items":{"$ref":"#/definitions/topicsTopic","type":"object"},"type":"array"}},"type":"object"},"topicsTopic":{"properties":{"createdAt":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"string"},"name":{"type":"string"},"taskIds":{"items":{"format":"int64","type":"string"},"type":"array"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"topicsTopicWithFetchedTasks":{"properties":{"createdAt":{"format":"date-time","type":"string"},"id":{"format":"int64","type":"string"},"name":{"type":"string"},"tasks":{"items":{"$ref":"#/definitions/tasksTask","type":"object"},"type":"array"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"topicsUpdateTopicsRequest":{"properties":{"topicsToUpdate":{"items":{"$ref":"#/definitions/topicsUpdateTopicsRequestTopicPrototype","type":"object"},"type":"array"}},"type":"object"},"topicsUpdateTopicsRequestTopicPrototype":{"properties":{"id":{"format":"int64","type":"string"},"name":{"type":"string"},"taskIds":{"items":{"format":"int64","type":"string"},"type":"array"}},"type":"object"},"topicsUpdateTopicsResponse":{"type":"object"}},"info":{"title":"api/topics/services.proto","version":"version not set"},"paths":{"/api/v1/change_board_order":{"post":{"operationId":"BoardService_ChangeBoardOrder","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsChangeBoardOrderRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsChangeBoardOrderResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/create_boards":{"post":{"operationId":"BoardService_CreateBoards","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsCreateBoardsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsCreateBoardsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/create_tasks":{"post":{"operationId":"TaskService_CreateTasks","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/tasksCreateTasksRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/tasksCreateTasksResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TaskService"]}},"/api/v1/create_topics":{"post":{"operationId":"TopicService_CreateTopics","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/topicsCreateTopicsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/topicsCreateTopicsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TopicService"]}},"/api/v1/delete_boards":{"post":{"operationId":"BoardService_DeleteBoards","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsDeleteBoardsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsDeleteBoardsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/delete_tasks":{"post":{"operationId":"TaskService_DeleteTasks","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/tasksDeleteTasksRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/tasksDeleteTasksResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TaskService"]}},"/api/v1/delete_topics":{"post":{"operationId":"TopicService_DeleteTopics","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/topicsDeleteTopicsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/topicsDeleteTopicsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TopicService"]}},"/api/v1/get_all_boards":{"post":{"operationId":"BoardService_GetAllBoards","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsGetAllBoardsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsGetAllBoardsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/get_all_dependencies":{"post":{"operationId":"TaskService_GetAllDependencies","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/tasksGetAllDependenciesRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/tasksGetAllDependenciesResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"summary":"Fetches all (direct and indirect) dependencies of tasks.","tags":["TaskService"]}},"/api/v1/get_board_content":{"post":{"operationId":"BoardService_GetBoardContent","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsGetBoardContentRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsGetBoardContentResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/get_board_order":{"post":{"operationId":"BoardService_GetBoardOrder","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsGetBoardOrderRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsGetBoardOrderResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/get_boards":{"post":{"operationId":"BoardService_GetBoards","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsGetBoardsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsGetBoardsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/get_tasks":{"post":{"operationId":"TaskService_GetTasks","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/tasksGetTasksRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/tasksGetTasksResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TaskService"]}},"/api/v1/get_topics":{"post":{"operationId":"TopicService_GetTopics","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/topicsGetTopicsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/topicsGetTopicsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TopicService"]}},"/api/v1/update_boards":{"post":{"operationId":"BoardService_UpdateBoards","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/boardsUpdateBoardsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/boardsUpdateBoardsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["BoardService"]}},"/api/v1/update_tasks":{"post":{"operationId":"TaskService_UpdateTasks","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/tasksUpdateTasksRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/tasksUpdateTasksResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TaskService"]}},"/api/v1/update_topics":{"post":{"operationId":"TopicService_UpdateTopics","parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/topicsUpdateTopicsRequest"}}],"responses":{"200":{"description":"A successful response.","schema":{"$ref":"#/definitions/topicsUpdateTopicsResponse"}},"default":{"description":"An unexpected error response.","schema":{"$ref":"#/definitions/rpcStatus"}}},"tags":["TopicService"]}}},"produces":["application/json"],"swagger":"2.0","tags":[{"name":"TopicService"}]}